refactor: call `getInputGroupCSS` to get the CSS style of an input group
