+++ packages/react-styled-core/src/Input/index.js
@@ -1,7 +1,7 @@
 import React, { forwardRef } from 'react';
 import InputBase from '../InputBase';
 import { useInputGroup } from '../InputGroup/context';
-import { useInputStyle } from './styles';
+import { getInputGroupCSS, useInputStyle } from './styles';
 
 const defaultSize = 'md';
 const defaultVariant = 'outline';
@@ -26,24 +26,6 @@ const Input = forwardRef((
     // - Fallback to the default value if the value is null or undefined
     size = (size ?? inputGroupSize) ?? defaultSize;
     variant = (variant ?? inputGroupVariant) ?? defaultVariant;
-
-    const useNegativeMargin = (variant === 'outline' || variant === 'filled');
-
-    css = {
-      '&:not(:first-child)': {
-        borderTopLeftRadius: 0,
-        borderBottomLeftRadius: 0,
-      },
-      '&:not(:last-child)': {
-        borderTopRightRadius: 0,
-        borderBottomRightRadius: 0,
-      },
-      // adjacent sibling
-      '&+&': {
-        marginLeft: useNegativeMargin ? -1 : 0,
-      },
-      ...css
-    };
   } else {
     // Use the default value if the value is null or undefined
     size = size ?? defaultSize;
@@ -56,7 +38,10 @@ const Input = forwardRef((
     <InputBase
       ref={ref}
       as="input"
-      css={css}
+      css={[ // TODO: combine the css array in v11
+        getInputGroupCSS({ variant }),
+        css,
+      ]}
       {...styleProps}
       {...rest}
     />
